j3dcore: use consistent locking when updating the renderAtoms array
You cannot lock updates to a member variable using itself if the reference is
ever changed.  The renderAtoms array in GeomtryAtom is resized when it does not
have enough room for all the open views.  This leads to crashes similar to the following:

Exception in thread "J3D-RenderStructureUpdateThread-5" java.lang.ArrayIndexOutOfBoundsException: 3
	at javax.media.j3d.GeometryAtom.getRenderAtom(GeometryAtom.java:241)
	at javax.media.j3d.RenderBin.processTransformChanged(RenderBin.java:3506)
	at javax.media.j3d.RenderBin.processMessages(RenderBin.java:1772)
	at javax.media.j3d.StructureUpdateThread.doWork(StructureUpdateThread.java:102)
	at javax.media.j3d.J3dThread.run(J3dThread.java:275)

Use the existing lockObj in this class to synchronize access.  This lock is already
being used to protect updates to the centroid update calculation, but both locked
sections are relatively short, if contention ends up being a problem, renderAtoms
could be changed to an ArrayList instead which can be resized.

Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>