j3dcore: avoid null pointer when reading raster depth-only from Canvas3d
Taken from java3d issue 593:

When I'm reading the z-buffer with a Raster, a NullPointerException occurs (in
fact, two). This is because the image raster is accessed whereas is not
initialized, and also because we send the depthComp to the native readRaster
function instead of the real buffer (i named it depthBuffer).

Also the destination offset is send to the native readRaster function instead of
source offset.

I made a very light patch that corrects this bugs.

Signed-off-by: Julian Ibarz <julian.ibarz@gmail.com>
Signed-off-by: Harvey Harrison <harvey.harrison@gmail.com>