Font3D: use the GeometryService as appropriate
This avoids a dependency on j3d-core-utils's com.sun.j3d.utils.geometry
package. If j3d-core-utils is present on the classpath, it can provide
the same backing implementation as before, but the option is now open
to provide an alternative service implementation if desired.

This addresses hharrison/java3d-core#17.