Pure Immediate Context usage fix
The Renderer now wraps a use/release context call pair around any of the
Pure Immediate mode operations in the main doWork loop.
The renderer also now lazily creates a context on the first Pure
Immediate mode call.
GarphicsContext3D calls makeCxtCurrent in doClear in the case where a
context had to be created (that creation call should in fact now be
redundant but it is left in)